--- a/compiler/fuzzer/src/input_pool.rs
+++ b/compiler/fuzzer/src/input_pool.rs
@@ -1,9 +1,9 @@
 use super::input::Input;
 use crate::{runner::RunResult, values::InputGeneration};
-use candy_vm::heap::{Heap, Text};
+use candy_vm::heap::{Heap, SymbolTable};
 use itertools::Itertools;
 use rand::{rngs::ThreadRng, seq::SliceRandom, Rng};
-use rustc_hash::{FxHashMap, FxHashSet};
+use rustc_hash::FxHashMap;
 use std::{cell::RefCell, rc::Rc};
 
 pub type Score = f64;
@@ -11,25 +11,16 @@ pub type Score = f64;
 pub struct InputPool {
     heap: Rc<RefCell<Heap>>,
     num_args: usize,
-    symbols: Vec<Text>,
+    symbol_table: SymbolTable,
     results_and_scores: FxHashMap<Input, (RunResult, Score)>,
 }
 
 impl InputPool {
-    pub fn new(num_args: usize, symbols_in_heap: &FxHashSet<Text>) -> Self {
-        let mut heap = Heap::default();
-
-        let mut symbols = symbols_in_heap
-            .iter()
-            .map(|symbol| symbol.clone_to_heap(&mut heap).try_into().unwrap())
-            .collect_vec();
-        symbols.push(Text::create(&mut heap, true, "True"));
-        symbols.push(Text::create(&mut heap, true, "False"));
-
+    pub fn new(num_args: usize, symbol_table: SymbolTable) -> Self {
         Self {
-            heap: Rc::new(RefCell::new(heap)),
+            heap: Rc::default(),
             num_args,
-            symbols,
+            symbol_table,
             results_and_scores: FxHashMap::default(),
         }
     }
@@ -46,7 +37,7 @@ impl InputPool {
         let mut rng = ThreadRng::default();
 
         if rng.gen_bool(0.1) || self.results_and_scores.len() < 20 {
-            return Input::generate(self.heap.clone(), self.num_args, &self.symbols);
+            return Input::generate(self.heap.clone(), self.num_args, &self.symbol_table);
         }
 
         let inputs_and_scores = self.results_and_scores.iter().collect_vec();
@@ -54,7 +45,7 @@ impl InputPool {
             .choose_weighted(&mut rng, |(_, (_, score))| *score)
             .unwrap();
         let mut input = (**input).clone();
-        input.mutate(&mut rng, &self.symbols);
+        input.mutate(&mut rng, &self.symbol_table);
         input
     }