--- a/compiler/language_server/src/debug_adapter/paused/stack_trace.rs
+++ b/compiler/language_server/src/debug_adapter/paused/stack_trace.rs
@@ -11,7 +11,7 @@ use crate::{
 use candy_frontend::{ast_to_hir::AstToHir, hir::Id, utils::AdjustCasingOfFirstLetter};
 use candy_vm::{
     fiber::FiberId,
-    heap::{Data, InlineObject},
+    heap::{Data, DisplayWithSymbolTable, InlineObject},
     lir::Lir,
     vm::Vm,
 };
@@ -129,7 +129,10 @@ impl PausedState {
                 );
                 (name, None, None)
             }
-            it => panic!("Unexpected callee: {it}"),
+            it => panic!(
+                "Unexpected callee: {}",
+                DisplayWithSymbolTable::to_string(&it, &lir.symbol_table),
+            ),
         };
         dap::types::StackFrame {
             id,