… compiler/language_server/src/features_candy/analyzer/module_analyzer.rs
… compiler/vm/benches/utils.rs
… compiler/vm/fuzz/fuzz_targets/vm.rs
… compiler/vm/src/builtin_functions.rs
… compiler/vm/src/fiber.rs
… compiler/vm/src/heap/mod.rs
… compiler/vm/src/heap/object.rs
… compiler/vm/src/heap/object_heap/function.rs
… compiler/vm/src/heap/object_heap/hir_id.rs
… compiler/vm/src/heap/object_heap/list.rs
… compiler/vm/src/heap/object_heap/mod.rs
… compiler/vm/src/heap/object_heap/struct_.rs
… compiler/vm/src/heap/object_heap/tag.rs
… compiler/vm/src/heap/object_inline/builtin.rs
… compiler/vm/src/heap/object_inline/int.rs
… compiler/vm/src/heap/object_inline/mod.rs
… compiler/vm/src/heap/object_inline/pointer.rs
… compiler/vm/src/heap/object_inline/port.rs
… compiler/vm/src/heap/object_representation.md
… compiler/vm/src/heap/symbol_table.rs
… compiler/vm/src/lib.rs
… compiler/vm/src/lir.rs
… compiler/vm/src/mir_to_lir.rs
… compiler/vm/src/tracer/stack_trace.rs
… compiler/vm/src/vm.rs
… packages/Examples/.goldens/_.candy/LIR.txt
… packages/Examples/.goldens/_package.candy/LIR.txt
… packages/Examples/.goldens/echo.candy/LIR.txt
… packages/Examples/.goldens/fibonacci.candy/LIR.txt
… packages/Examples/.goldens/helloWorld.candy/LIR.txt