**Depends on #583**

<!-- Please enter the corresponding issue ID: -->


<!-- Please summarize your changes: -->
Intern symbols in a symbol table instead of allocating them over and over again for new tags. Unfortunately, the code got more verbose because the symbol table has to be passed around quite a few places, which is also why I had to create new traits `OrdWithSymbolTable` and `DisplayWithSymbolTable`.

## Benchmark Results


`fibonacci 20` instead of the default `fibonacci 15`

```
Before: cargo run --release -- run packages/examples/fibonacci.candy
  Time (mean ± σ):      1.158 s ±  0.064 s    [User: 1.122 s, System: 0.037 s]
  Range (min … max):    1.066 s …  1.256 s    10 runs

After: cargo run --release -- run packages/examples/fibonacci.candy
  Time (mean ± σ):     970.6 ms ±  45.0 ms    [User: 934.7 ms, System: 37.3 ms]
  Range (min … max):   874.2 ms … 1035.2 ms    10 runs
```

CI Results: https://github.com/candy-lang/candy/commit/b32c6b49f51d8c7258826b71be4cd6290b1548b3#commitcomment-122887486


### Checklist
<!-- Please check if your PR fulfills the following requirements: -->

- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)